fix(utils): include PR feedback
