+++ packages/utils/src/lib/case-conversions.unit.test.ts
@@ -1,9 +1,25 @@
+import { describe, expect, it } from 'vitest';
 import {
   camelCaseToKebabCase,
+  capitalize,
   kebabCaseToCamelCase,
   toSentenceCase,
   toTitleCase,
 } from './case-conversions.js';
+
+describe('capitalize', () => {
+  it('should transform the first string letter to upper case', () => {
+    expect(capitalize('code PushUp')).toBe('Code PushUp');
+  });
+
+  it('should leave the first string letter in upper case', () => {
+    expect(capitalize('Code PushUp')).toBe('Code PushUp');
+  });
+
+  it('should accept empty string', () => {
+    expect(capitalize('')).toBe('');
+  });
+});
 
 describe('kebabCaseToCamelCase', () => {
   it('should convert simple kebab-case to camelCase', () => {
