+++ packages/utils/src/lib/text-formats/table.ts
@@ -5,7 +5,7 @@ import type {
   TableColumnObject,
   TableColumnPrimitive,
 } from '@code-pushup/models';
-import { capitalize } from '../transform.js';
+import { capitalize } from '@code-pushup/utils';
 
 export function rowToStringArray({ rows, columns = [] }: Table): string[][] {
   if (Array.isArray(rows.at(0)) && typeof columns.at(0) === 'object') {
